ADD export_data.py and UPDATE team_code.py
Implements basic export functionality to export patient metadata and recordings data into csv files.
The directory structure is the same as in the original data. The get_features function in team_code.py has been updated to make the required data more easily accessible.

The patient meta data is exported from a numpy array and includes the quality score - it's probably better to rewrite this.